+++ src/entrypoints/ui/utils/tempad.ts
@@ -17,7 +17,8 @@ const LIB_DISPLAY_NAMES = {
   '@baidu/one-ui-pro': 'ONE UI Pro',
   '@baidu/one-charts': 'ONE Charts',
   '@baidu/light-ai-react': 'Light AI',
-  'dls-icons-react': 'DLS Icons'
+  'dls-icons-react': 'DLS Icons',
+  'dls-illustrations-react': 'DLS Illus.'
 } as Record<string, string>
 
 export function getTemPadComponent(node: SceneNode): TemPadComponent | null {
@@ -28,6 +29,22 @@ export function getTemPadComponent(node: SceneNode): TemPadComponent | null {
   const tempadData = JSON.parse(
     node.getSharedPluginData(NS, 'source') || 'null'
   ) as TemPadSource | null
+
+  if (
+    tempadData?.libName === '@baidu/one-ui' &&
+    (tempadData?.name === 'Icon' || tempadData?.name === 'Illustration')
+  ) {
+    const tree = JSON.parse(node.getSharedPluginData(NS, 'tree') || 'null')
+    try {
+      const iconNode = tree.slots.default.children[0]
+      tempadData.libName =
+        tempadData.name === 'Illustration' ? 'dls-illustrations-react' : iconNode.props.libName.v
+      tempadData.name = iconNode.props.name.v
+    } catch (e) {
+      console.error(e)
+    }
+  }
+
   const libDisplayName = tempadData?.libName ? LIB_DISPLAY_NAMES[tempadData.libName] : null
 
   const codeNode = node.findChild((n) => n.type === 'TEXT' && n.name === '代码') as TextNode